requirements.txt
tkinter
ttkthemes
tkcalendar
requests
pandas
python-dotenv
openpyxl
matplotlib